Use a common module to avoid code duplication
