+++ amq-quarkus-school-timetabling/README.adoc
@@ -114,22 +114,7 @@ $ ./target/*-runner
 [source, shell]
 ----
 $ cd amq-quarkus-school-timetabling-client
-$ mvn package -Dnative -DskipTests -Dquarkus.profile=native
-----
-+
-[WARNING]
-====
-The `-Dquarkus.profile=native` is a temporary workaround for https://github.com/quarkusio/quarkus/issues/13341[an issue in Quarkus 1.10].
-====
-
-. Run a database. By default, `application.properties` is configured for H2:
-.. Download the http://www.h2database.com/html/download.html[H2 database] (Platform-independent zip) and unzip it.
-.. Start the H2 server with the option `-ifNotExists` (not recommended in production):
-+
-[source, shell]
-----
-$ cd h2/bin
-$ java -cp h2*.jar org.h2.tools.Server -ifNotExists
+$ mvn package -Dnative -DskipTests
 ----
 
 . Run the native executable:
