+++ amq-quarkus-school-timetabling/amq-quarkus-school-timetabling-client/pom.xml
@@ -15,29 +15,23 @@
 
   <dependencies>
     <dependency>
+      <groupId>org.acme</groupId>
+      <artifactId>amq-quarkus-school-timetabling-common</artifactId>
+      <exclusions>
+        <!-- Avoid bootstrapping the optaplanner-quarkus extension. -->
+        <exclusion>
+          <groupId>org.optaplanner</groupId>
+          <artifactId>optaplanner-quarkus</artifactId>
+        </exclusion>
+      </exclusions>
+    </dependency>
+    <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-resteasy</artifactId>
     </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-resteasy-jackson</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.quarkus</groupId>
-      <artifactId>quarkus-hibernate-orm-panache</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.quarkus</groupId>
-      <artifactId>quarkus-jdbc-h2</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.quarkus</groupId>
-      <artifactId>quarkus-hibernate-orm-rest-data-panache</artifactId>
-    </dependency>
-    <!-- OptaPlanner is required due to usage of Score. -->
-    <dependency>
-      <groupId>org.optaplanner</groupId>
-      <artifactId>optaplanner-quarkus</artifactId>
     </dependency>
     <dependency>
       <groupId>org.optaplanner</groupId>
@@ -57,11 +51,6 @@
     <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-junit5</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.quarkus</groupId>
-      <artifactId>quarkus-test-h2</artifactId>
       <scope>test</scope>
     </dependency>
     <dependency>
