--- amq-quarkus-school-timetabling/amq-quarkus-school-timetabling-client/src/main/java/org/acme/schooltimetabling/domain/TimeTable.java
@@ -1,88 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.acme.schooltimetabling.domain;
-
-import java.util.List;
-
-import org.optaplanner.core.api.score.buildin.hardsoft.HardSoftScore;
-import org.optaplanner.core.api.solver.SolverStatus;
-
-public class TimeTable {
-
-    private List<Timeslot> timeslotList;
-    private List<Room> roomList;
-    private List<Lesson> lessonList;
-
-    private HardSoftScore score;
-
-    // Ignored by OptaPlanner, used by the UI to display solve or stop solving button
-    private SolverStatus solverStatus;
-
-    // No-arg constructor required for OptaPlanner
-    public TimeTable() {
-    }
-
-    public TimeTable(List<Timeslot> timeslotList, List<Room> roomList, List<Lesson> lessonList) {
-        this.timeslotList = timeslotList;
-        this.roomList = roomList;
-        this.lessonList = lessonList;
-    }
-
-    // ************************************************************************
-    // Getters and setters
-    // ************************************************************************
-
-    public List<Timeslot> getTimeslotList() {
-        return timeslotList;
-    }
-
-    public List<Room> getRoomList() {
-        return roomList;
-    }
-
-    public List<Lesson> getLessonList() {
-        return lessonList;
-    }
-
-    public HardSoftScore getScore() {
-        return score;
-    }
-
-    public SolverStatus getSolverStatus() {
-        return solverStatus;
-    }
-
-    public void setTimeslotList(List<Timeslot> timeslotList) {
-        this.timeslotList = timeslotList;
-    }
-
-    public void setRoomList(List<Room> roomList) {
-        this.roomList = roomList;
-    }
-
-    public void setLessonList(List<Lesson> lessonList) {
-        this.lessonList = lessonList;
-    }
-
-    public void setScore(HardSoftScore score) {
-        this.score = score;
-    }
-
-    public void setSolverStatus(SolverStatus solverStatus) {
-        this.solverStatus = solverStatus;
-    }
-}
