+++ amq-quarkus-school-timetabling/amq-quarkus-school-timetabling-client/src/main/resources/application.properties
@@ -22,15 +22,6 @@
 # timeTable.demoData=LARGE
 
 ########################
-# Database properties
-########################
-
-# "jdbc:h2:mem" doesn't work in native mode, but native mode uses %prod properties
-quarkus.datasource.db-kind=h2
-quarkus.datasource.jdbc.url=jdbc:h2:mem:school-timetabling;DB_CLOSE_DELAY=-1
-quarkus.hibernate-orm.database.generation=drop-and-create
-
-########################
 # Messaging
 ########################
 
@@ -44,13 +35,6 @@ mp.messaging.outgoing.solver_request.durable=true
 # Configure the Kafka source (we read from it)
 mp.messaging.incoming.solver_response.connector=smallrye-amqp
 mp.messaging.incoming.solver_response.durable=true
-
-########################
-# Native overrides
-########################
-
-# In pom.xml, the "native" maven profile triggers the "native" quarkus profile.
-%native.quarkus.datasource.jdbc.url=jdbc:h2:tcp://localhost/mem:school-timetabling
 
 # TODO: Remove after switching to fast jar by default.
 quarkus.package.type=fast-jar
