+++ amq-quarkus-school-timetabling/amq-quarkus-school-timetabling-client/src/test/java/org/acme/schooltimetabling/TimeTableResourceTest.java
@@ -33,8 +33,8 @@ import org.acme.schooltimetabling.domain.Lesson;
 import org.acme.schooltimetabling.domain.Room;
 import org.acme.schooltimetabling.domain.TimeTable;
 import org.acme.schooltimetabling.domain.Timeslot;
-import org.acme.schooltimetabling.messaging.SolverRequest;
-import org.acme.schooltimetabling.messaging.SolverResponse;
+import org.acme.schooltimetabling.message.SolverRequest;
+import org.acme.schooltimetabling.message.SolverResponse;
 import org.acme.schooltimetabling.persistence.TimeTableRepository;
 import org.acme.schooltimetabling.rest.TimeTableResource;
 import org.eclipse.microprofile.reactive.messaging.Message;
