+++ amq-quarkus-school-timetabling/amq-quarkus-school-timetabling-common/.gitignore
@@ -1,0 +1,13 @@
+/target
+/build
+/local
+
+# Eclipse, Netbeans and IntelliJ files
+/.*
+!.gitignore
+!.dockerignore
+!.mvn
+/nbproject
+/*.ipr
+/*.iws
+/*.iml
