+++ amq-quarkus-school-timetabling/amq-quarkus-school-timetabling-common/src/main/java/org/acme/schooltimetabling/domain/Timeslot.java
@@ -19,22 +19,26 @@ package org.acme.schooltimetabling.domain;
 import java.time.DayOfWeek;
 import java.time.LocalTime;
 
+import org.optaplanner.core.api.domain.lookup.PlanningId;
+
 public class Timeslot {
 
+    @PlanningId
     private Long id;
+
     private DayOfWeek dayOfWeek;
     private LocalTime startTime;
     private LocalTime endTime;
 
-    // No-arg constructor required for Hibernate
-    public Timeslot() {
+    Timeslot() {
+        // Required for JSON deserialization.
     }
 
     public Timeslot(long id, DayOfWeek dayOfWeek, LocalTime startTime, LocalTime endTime) {
-        this.id = id;
         this.dayOfWeek = dayOfWeek;
         this.startTime = startTime;
         this.endTime = endTime;
+        this.id = id;
     }
 
     @Override
