+++ amq-quarkus-school-timetabling/amq-quarkus-school-timetabling-common/src/main/java/org/acme/schooltimetabling/message/SolverRequest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.acme.schooltimetabling.messaging;
+package org.acme.schooltimetabling.message;
 
 import org.acme.schooltimetabling.domain.TimeTable;
