--- amq-quarkus-school-timetabling/amq-quarkus-school-timetabling-solver/src/main/java/org/acme/schooltimetabling/messaging/ErrorInfo.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.acme.schooltimetabling.messaging;
-
-public class ErrorInfo {
-    private String exceptionClassName;
-    private String exceptionMessage;
-
-    ErrorInfo() {
-        // Required for JSON deserialization.
-    }
-
-    public ErrorInfo(String exceptionClassName, String exceptionMessage) {
-        this.exceptionClassName = exceptionClassName;
-        this.exceptionMessage = exceptionMessage;
-    }
-
-    public String getExceptionClassName() {
-        return exceptionClassName;
-    }
-
-    public String getExceptionMessage() {
-        return exceptionMessage;
-    }
-}
